Checks if wordlist exists
This just checks for the hard-coded word-list.

Note: 
It would probably be better if the user provides the word-list. You could use the "-w" flag (-w wordlist.txt) for example. If the user provides no word-list then the script can use a hard-coded one (like the one being used now).